Simplify and consolidate target framework
